Refactor `SingleSubredditViewController` in preparation for prefetching implementation
