+++ Slide for Reddit/UIApplication+Extensions.swift
@@ -1,0 +1,17 @@
+//
+//  UIApplication+Extensions.swift
+//  Slide for Reddit
+//
+//  Created by Jonathan Cole on 7/6/18.
+//  Copyright © 2018 Haptic Apps. All rights reserved.
+//
+
+import Foundation
+
+extension UIApplication {
+
+    var statusBarView: UIView? {
+        return value(forKey: "statusBar") as? UIView
+    }
+    
+}
