+++ Slide for Reddit/UIButton+Extensions.swift
@@ -1,0 +1,26 @@
+//
+//  UIButton+Extensions.swift
+//  Slide for Reddit
+//
+//  Created by Jonathan Cole on 7/6/18.
+//  Copyright © 2018 Haptic Apps. All rights reserved.
+//
+
+import UIKit
+
+//https://stackoverflow.com/a/50127204/3697225
+extension UIButton {
+    func leftImage(image: UIImage, renderMode: UIImageRenderingMode) {
+        self.setImage(image.withRenderingMode(renderMode), for: .normal)
+        self.imageEdgeInsets = UIEdgeInsets(top: 0, left: image.size.width / 2, bottom: 0, right: image.size.width / 2)
+        self.contentHorizontalAlignment = .left
+        self.imageView?.contentMode = .scaleAspectFit
+    }
+
+    func rightImage(image: UIImage, renderMode: UIImageRenderingMode) {
+        self.setImage(image.withRenderingMode(renderMode), for: .normal)
+        self.imageEdgeInsets = UIEdgeInsets(top: 0, left: image.size.width / 2, bottom: 0, right: 0)
+        self.contentHorizontalAlignment = .right
+        self.imageView?.contentMode = .scaleAspectFit
+    }
+}
