+++ Slide for Reddit/UIImage+Extensions.swift
@@ -51,4 +51,39 @@ extension UIImage {
     func menuIcon() -> UIImage {
         return self.getCopy(withSize: CGSize(width: 20, height: 20), withColor: ColorUtil.fontColor)
     }
+
+    func cropToBounds(image: UIImage, width: Double, height: Double) -> UIImage {
+
+        let contextImage: UIImage = UIImage(cgImage: image.cgImage!)
+
+        let contextSize: CGSize = contextImage.size
+
+        var posX: CGFloat = 0.0
+        var posY: CGFloat = 0.0
+        var cgwidth: CGFloat = CGFloat(width)
+        var cgheight: CGFloat = CGFloat(height)
+
+        // See what size is longer and create the center off of that
+        if contextSize.width > contextSize.height {
+            posX = ((contextSize.width - contextSize.height) / 2)
+            posY = 0
+            cgwidth = contextSize.height
+            cgheight = contextSize.height
+        } else {
+            posX = 0
+            posY = ((contextSize.height - contextSize.width) / 2)
+            cgwidth = contextSize.width
+            cgheight = contextSize.width
+        }
+
+        let rect: CGRect = CGRect.init(x: posX, y: posY, width: cgwidth, height: cgheight)
+
+        // Create bitmap image from context using the rect
+        let imageRef: CGImage = (contextImage.cgImage?.cropping(to: rect)!)!
+
+        // Create a new image based on the imageRef and rotate back to the original orientation
+        let image: UIImage = UIImage.init(cgImage: imageRef, scale: image.scale, orientation: image.imageOrientation)
+
+        return image
+    }
 }
