+++ Slide for Reddit/UINavigationController+Extensions.swift
@@ -1,0 +1,15 @@
+//
+//  UINavigationController+Extensions.swift
+//  Slide for Reddit
+//
+//  Created by Jonathan Cole on 7/6/18.
+//  Copyright © 2018 Haptic Apps. All rights reserved.
+//
+
+import UIKit
+
+extension UINavigationController {
+    override func topMostViewController() -> UIViewController {
+        return self.visibleViewController!.topMostViewController()
+    }
+}
